Moved all classes into the package.
