+++ src/edu/cpp/cs/cs141/GECLYfinalproj/Main.java
@@ -1,10 +1,12 @@
+package edu.cpp.cs.cs141.GECLYfinalproj;
+
 /**
  * CS 141: Intro to Programming and Problem Solving
  * Professor: Edwin Rodríguez
  *
  * Programming Final Project
  *
- * Ninja-Assassin Game
+ * edu.cpp.cs.cs141.GECLYfinalproj.Ninja-Assassin Game
  *
  * Team GECLY Info:
  * Gavin Kremer
@@ -14,6 +16,11 @@
  * Yan Huang (Lilli)
  *
  */
-public class ConsoleInterface extends UserInterface{
+public class Main {
+
+	public static void main(String[] args) {
+		System.out.println("Hello");
+	}
 
 }
+
